requirements.txt
# Core dependencies
autogen>=0.2.0
ag2[openai]>=0.2.0
openai>=0.27.0  # For OpenAI API access
requests>=2.28.0  # For HTTP requests
typing>=3.7.4

# Development dependencies
pytest>=7.0.0  # For testing
black>=22.0.0  # For code formatting
flake8>=4.0.0  # For linting
mypy>=0.910  # For type checking

# Optional dependencies
tqdm>=4.65.0  # For progress bars
python-dotenv>=1.0.0  # For environment variable management